chore: 기존 토글을 다크 모드 토글로 변경 및 로직 추가
